Refactored server to mjs file to use es6 with node.
